Rename group state change/update trigger to member state change/update.
This does reflect real intention of these triggers and it is less confusing than refering a group which state can stay intact.
Group, when it changes or receives an update, does trigger rules just as an item hence does not need to be mentioned directly.

Additional changes include alignment of naming conventions, extraction of context interfaces and basic unit tests for things actions.

Signed-off-by: Łukasz Dywicki <luke@code-house.org>