Add support jrxtx serial library and bridges within wmbus and mbus bindings.
The jrxtx library is another flavor of forked rxtx code.
It seems to be needed in order to limit discarted data by jmbus (fixes #28).

The way of work is embedding of native parts coming from nrjavaserial.
This little trick allows to keep jrxtx api but stick with supported version of rxtx fork.

Signed-off-by: Łukasz Dywicki <luke@code-house.org>